Set Smack ambient to match run label.
Set the Smack networking ambient label to match the
run label of systemd. System services may expect to
communicate with external services over IP. Setting
the ambient label assigns that label to IP packets
that do not include CIPSO headers. This allows systemd
and the services it spawns access to unlabeled IP
packets, and hence external services.

A system may choose to restrict network access to
particular services later in the startup process.
This is easily done by resetting the ambient label
elsewhere.

Change-Id: Ia9c8d8744b732b2ee1126d2e446585d16fa56908
Signed-off-by: Casey Schaufler <casey.schaufler@intel.com>