execute: add SmackExecLabel key for services
At the filesystem level, setting the runtime Smack label for an
executable is possible by setting the SMACK64EXEC security extended
attribute on the file, but if the runtime Smack label should depend on
the execution environment, the process needs to change its own Smack
label.

This commit introduces a new SmackExecLabel key for this purpose. For
example, 'systemd --user' may need to run with a different label as
compared to 'systemd --system', so SmackExecLabel could be set in
user@.service.

Setting the label will fail if the service lacks capabilities to set its
own Smack label.

This code has been rebased from systemd v208 and adapted to a new
bus code.

Change-Id: I8eb8eb222342c9d0ae7e28530fe7e4d51493b83e
Signed-off-by: Patrick McCarty <patrick.mccarty@linux.intel.com>
Signed-off-by: Łukasz Stelmach <l.stelmach@samsung.com>