journald: check if socket is open before forwarding to syslog
In journald messages are forwarded to syslog using sendmsg() system call.
According to documantation, the send() call may be used only when the socket
is in a connected state. Check the state of the socket before sending message.

This patch causes about 35% decrease in CPU usage on Trats2 when ForwardToSyslog=yes
but the syslog socket is not open.

Change-Id: I7725ee7867350fa59db3734c474a61140e549754
Signed-off-by: Piotr Wilczek <p.wilczek@samsung.com>