Add boost shutting down feature. This feature is to shorten tizen shutting down time usage into 1 second. And this version of commit is the draft prototype, which can make Tizen shutting down in 866ms.
The original shutting procedure requires to build all the shutdown.target dependency and stop all of the related services recursively. With experiments, we found the major time wasting is when systemd builds the dependency transaction and to stops all of the unnecessary service.

Hence, I implement 'boost' shutting down feature. Unlike 'systemctl -f poweroff', I will run the commands which defined in system/shutdown.wants/*.services, and then call the same procedurer of 'systemctl -f poweroff'.

The modifications are in two parts:
#1. add 'boost' option(including dbus message) in src/core/dbus-manager.c, src/core/shutdown.c, src/login/org.freedesktop.login1.conf and src/systemctl/systemctl.c.

#2. handle shutdown.boost file and run the commands parallel in boost-shutdown.c and boost-shutdown.h. A cmd_queue is added in systemd manager which will read the commands when tizen boot up. And the commands inside manager->cmd_queue will execute parallel at the beginning of power off.

Change-Id: Ib71d93cbb40ebe9e26190fbdb3b459a89b2435ae