smack: Don't follow symbolic links for setting smack labels in /dev
When setting the smack label on a symbolic link, the link target,
instead of the link itself was updated. This change correctly sets
the symbolic link label as looks like is intended.

Also this fix is included in the latest systemd implementation in
upstream, coming from 5dfc54615a1eacea18106383c964425cebd67c30 in
git://anongit.freedesktop.org/systemd/systemd, athough it is included
with a number of other changes that are not easily backported.

Change-Id: I1aeef3caad73ccc20efd1f16019740f8a415fde3
Signed-off-by: Kazunori Kobayashi <kkobayas@igel.co.jp>