core: rework device state logic
This change introduces a new state "tentative" for device units. Device
units are considered "plugged" when udev announced them, "dead" when
they are not available in the kernel, and "tentative" when they are
referenced in /proc/self/mountinfo or /proc/swaps but not (yet)
announced via udev.

This should fix a race when device nodes (like loop devices) are created
and immediately mounted. Previously, systemd might end up seeing the
mount unit before the device, and would thus pull down the mount because
its BindTo dependency on the device would not be fulfilled.

Change-Id: Ifb1bc57fa4fdc657db575c51eddb621919ac1fb9
Signed-off-by: Maciej Wereski <m.wereski@partner.samsung.com>
Backported-by: Maciej Wereski <m.wereski@partner.samsung.com>