core: don't change removed devices to state "tentative"
Commit 628c89c introduced the "tentative" device state, which caused
devices to go from "plugged" to "tentative" on a remove uevent. This
breaks the cleanup of stale mounts (see commit 3b48ce4), as that only
applies to "dead" devices.

The "tentative" state only really makes sense on adding a device when
we don't know where it was coming from (i. e. not from udev). But when
we get a device removal from udev we definitively know that it's gone,
so change the device state back to "dead" as before 628c89c.

Change-Id: I3987416c337e0b827d2d8843e41d66bf129c4d2a
Signed-off-by: Maciej Wereski <m.wereski@partner.samsung.com>
Backported-by: Maciej Wereski <m.wereski@partner.samsung.com>