core: fix the CGROUP spawning error as a workaround
When logging in and out continually, the below error occurs.

* systemd: Failed at step CGROUP spawning /usr/lib/systemd/systemd: No
such file or directory

This is mainly because the cgroup path of systemd user session does not
exists even though that cgroup is marked as 'realized'. That is
definitely bug and it is already reported into systemd mainline last
January. But there is no _right_ solution right now and it looks like a
picky problem to resolve. So I made this patch as a workaround for Tizen
TDC demonstration.

Change-Id: Ie2e164a4e4daeb88fc102e5fa88e0faca28088b0
Signed-off-by: Sangjung Woo <sangjung.woo@samsung.com>