core: Fix path for opening ffs endpoint ep0
ffs_address_create() expects an absolute path to the file that is
supposed to be opened. The path specified only leads to the directory
containing the endpoint ep0 not the endpoint itself. This commit adds
the endpoints name to the path.

origin: https://github.com/systemd/systemd/commit/27a6ea9163

Change-Id: I0411df2a1f99b19edd121c9a3ed97b0966c8991a
Signed-off-by: Georgia Brikis <g.brikis@samsung.com>